FEM: elmer, implement generation from FC analysis
Initial implementation of case generation from the FreeCAD analysis.
Support includes the analysis types static, frequency and thermomech
with the following constraints: fixed, force, displacement, self weight,
temperature and initial temperature. Material object is honored but
support for multiple materials is still missing. Also missing is post
processing.

The log of the execution of ElmerSolver is exposed via a TextDocument
named ElmerSolverOutput. This object is created/updated every time the
solver is executed.